Keep Coolify runtime deploy test aligned
